기능 추가)몬스터 공격 소환 스텟
몬스터 스텟설정 및 소환 공격

+++ Client/Code/CUI_HpBar.cpp
@@ -153,7 +153,7 @@ void CUI_HpBar::Free()
 {
 }
 
-void CUI_HpBar::Get_Object_Transform(Engine::CTransform * vTransform, float* HP)
+void CUI_HpBar::Get_Object_Transform(Engine::CTransform * vTransform, UINT* HP)
 {
 	m_pObjectTransform = vTransform;
 	m_fObjectHp = HP;
